Added Radian class

+++ src/Math.cpp
@@ -3,10 +3,12 @@
 
 namespace core {
 
-const float Math::PI = float( 4.0 * atan( 1.0f ) );
-const float Math::TWO_PI = float( 2.0 * PI );
-const float Math::TAU = float( 2.0 * PI );
-const float Math::HALF_PI = float( 0.5 * PI );
-const float Math::LOG2 = std::log(float(2.0));
+const float Math::PI = 4.0f * atan(1.0f);
+const float Math::TWO_PI = 2.0f * PI;
+const float Math::TAU = 2.0f * PI;
+const float Math::HALF_PI = 0.5f * PI;
+const float Math::LOG2 = std::log(2.0f);
+
+
 
 } // namespace
